refactor: move utility types to internals
