--- packages/pure-parse/src/parsers/object/index.ts
@@ -1,1 +1,0 @@
-export * from './object'
